move restaurant mapper to ui state file
